Integrate Linting Checks via GitHub Actions and `lintr`
Integrated `usethis::use_github_action("lint")` to set up a GitHub Action for linting the codebase, ensuring adherence to a consistent coding style. Additionally, added an entry for `lintr::lint_package()` on the development scripts as an iterated command to be run before each commit, aiding in early identification and rectification of potential coding issues. This combination of automated and manual linting processes will help maintain code quality and readability as the project evolves.